requirements.txt
pandas
requests
beautifulsoup4
zstandard
sqlalchemy
python-chess